requirements.txt
git+https://github.com/Rapptz/discord.py
python-dotenv==0.20.0
colorama==0.4.5
requests==2.28.1
aiofiles==0.8.0
httpx==0.23.0